# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.00"
    },
    "paths": {
...
        "/api/cars": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get cars list",
                "operationId": "cars",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Page items limit"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/BaseResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/PaginateResponse"
                                        },
                                        {
                                            "properties": {
                                                "result": {
                                                    "properties": {
                                                        "cars": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/components/schemas/CarResource"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/cars/{car_uuid}": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get car by uuid",
                "operationId": "car_by_uuid",
                "parameters": [
                    {
                        "name": "car_uuid",
                        "in": "path",
                        "description": "Car uuid",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/BaseResponse"
                                        },
                                        {
                                            "properties": {
                                                "result": {
                                                    "properties": {
                                                        "car": {
                                                            "$ref": "#/components/schemas/CarResource"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/cars/{car_uuid}/lock": {
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Lock car by uuid",
                "operationId": "car_lock_by_uuid",
                "parameters": [
                    {
                        "name": "car_uuid",
                        "in": "path",
                        "description": "Car uuid",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/BaseResponse"
                                        },
                                        {
                                            "properties": {
                                                "result": {
                                                    "properties": {
                                                        "status": {
                                                            "type": "boolean",
                                                            "example": "true"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Can't lock car"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/cars/{car_uuid}/unlock": {
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Unlock car by uuid",
                "operationId": "car_unlock_by_uuid",
                "parameters": [
                    {
                        "name": "car_uuid",
                        "in": "path",
                        "description": "Car uuid",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/BaseResponse"
                                        },
                                        {
                                            "properties": {
                                                "result": {
                                                    "properties": {
                                                        "status": {
                                                            "type": "boolean",
                                                            "example": "true"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Can't access car to unlock"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CarResource": {
                "title": "Car resource",
                "properties": {
                    "uuid": {
                        "schema": "CarResource",
                        "title": "UUID",
                        "description": "Car UUID",
                        "type": "string",
                        "example": "d1b0c6be-e64d-4baf-acec-17e7c6162216"
                    },
                    "company": {
                        "schema": "CarResource",
                        "title": "Company name",
                        "description": "Car company name",
                        "type": "string",
                        "example": "BMW"
                    },
                    "model_family": {
                        "schema": "CarResource",
                        "title": "Model family",
                        "description": "Car model family",
                        "type": "string",
                        "example": "X"
                    },
                    "model_number": {
                        "schema": "CarResource",
                        "title": "Model number",
                        "description": "Car model number",
                        "type": "string",
                        "example": "100"
                    },
                    "state": {
                        "schema": "CarResource",
                        "title": "State",
                        "description": "Car state",
                        "type": "string",
                        "enum": [
                            "free",
                            "busy"
                        ],
                        "example": "free"
                    }
                },
                "type": "object"
            },
            "BaseResponse": {
                "description": "base response",
                "properties": {
                    "result": {
                        "title": "Result object",
                        "type": "object"
                    },
                    "message": {
                        "title": "Error message",
                        "type": "string"
                    },
                    "errors": {
                        "title": "Errors array",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
            "PaginateResponse": {
                "description": "paginate response",
                "properties": {
                    "result": {
                        "title": "Result object",
                        "properties": {
                            "total": {
                                "type": "integer",
                                "example": "10"
                            },
                            "limit": {
                                "type": "integer",
                                "example": "10"
                            },
                            "page": {
                                "type": "integer",
                                "example": "1"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "access_token for access to /user and /cars",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}